nfacct: change prototype of nfacct_snprintf() to support output in XML
This patch adds supports for XML output for nfacct_snprintf().

This also bumps the library current version number and it adds
a new version in the symbol map. I decided to do so because,
although no official release has happened so far. Many distributors
have already decided to provide packages for this.

We don't provide the old symbol version for nfacct_snprintf,
I think that's too much (given the lifetime of it). But we make
sure that we break the binary gracefully in case of version
mismatches (not silently, with unexpectable side effects).

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>